requirements.txt
django
djangorestframework
psycopg2-binary
numpy
pandas
scikit-learn
tensorflow
torch
transformers
spacy
celery
requests
